requirements.txt
pandas
matplotlib
tensorflow-gpu
numpy
scikit-learn